[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8706 Former-commit-id: 831412e0be10f3316166f6f2805d9af9422635f7

+++ icu4c/source/extra/uconv/uconv.cpp
@@ -595,7 +595,7 @@ static UBool convertFile(const char *pname,
         if (flush && !willexit && cbufp != (buf + rd)) {
             char pos[32];
             sprintf(pos, "%u", infoffset);
-            UnicodeString str(pos, (uprv_strlen(pos) + 1));
+            UnicodeString str(pos, (int32_t)(uprv_strlen(pos) + 1));
             initMsg(pname);
             u_wmsg(stderr, "premEndInput", str.getBuffer());
             willexit = 1;
